server/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-bcrypt
flask-jwt-extended
python-dotenv
Pillow
pdf2image
opencv-python
numpy
pdfplumber
paddlepaddle
paddleocr
werkzeug
flask-mail
